+++ src/main/java/net/geforcemods/securitycraft/items/AdminToolItem.java
@@ -8,6 +8,7 @@ import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.api.IModuleInventory;
 import net.geforcemods.securitycraft.api.IOwnable;
 import net.geforcemods.securitycraft.api.IPasswordProtected;
+import net.geforcemods.securitycraft.blockentities.DisplayCaseBlockEntity;
 import net.geforcemods.securitycraft.blockentities.SecretSignBlockEntity;
 import net.geforcemods.securitycraft.misc.ModuleType;
 import net.geforcemods.securitycraft.util.PlayerUtils;
@@ -49,6 +50,9 @@ public class AdminToolItem extends Item {
 			TileEntity te = world.getTileEntity(pos);
 
 			if (te != null) {
+				if (te instanceof DisplayCaseBlockEntity && (((DisplayCaseBlockEntity) te).isOpen() && ((DisplayCaseBlockEntity) te).getDisplayedStack().isEmpty()))
+					return EnumActionResult.PASS;
+
 				boolean hasInfo = false;
 
 				if (te instanceof IOwnable) {